added navbar with darktheme
