+++ frontend/src/App.js
@@ -1,24 +1,14 @@
 import React, { useState } from 'react';
-//import LinkForm from './components/LinkForm';
-import { ThemeProvider } from 'styled-components';
-import { GlobalStyles } from './components/globalStyles';
-import { lightTheme, darkTheme } from './components/Themes';
 import './App.css';
 import HomeScreen from './screens/Home';
 import ButtonAppBar from './components/navBar';
 
 const App = () => {
-  const [theme, setTheme] = useState('light');
-  const themeToggler = () => {
-    theme === 'light' ? setTheme('dark') : setTheme('light');
-  };
-
   return (
-    <ThemeProvider theme={theme === 'light' ? lightTheme : darkTheme}>
-      <GlobalStyles />
+    <div className="App">
       <ButtonAppBar />
       <HomeScreen />
-    </ThemeProvider>
+    </div>
   );
 };
 //<span className="PoweredByDscTiet">Powered By DSC TIET</span>
